// src/app/core/enum/meal.combination.enum.ts
export enum MealCombinationEnum {
    TWO_MEALS = '2 meals',
    THREE_MEALS = '3 meals',
    FOUR_MEALS = '4 meals',
    FIVE_MEALS = '5 meals',
    SIX_MEALS = '6 meals',
    SEVEN_MEALS = '7 meals',
    EIGHT_MEALS = '8 meals',
    NINE_MEALS = '9 meals',
    TEN_MEALS = '10 meals',
    TWO_MEALS_ONE_SNACK = '2 meals + snack',
    THREE_MEALS_ONE_SNACK = '3 meals + snack',
    FOUR_MEALS_ONE_SNACK = '4 meals + snack',
    FIVE_MEALS_ONE_SNACK = '5 meals + snack',
    SIX_MEALS_ONE_SNACK = '6 meals + snack',
    SEVEN_MEALS_ONE_SNACK = '7 meals + snack',
    EIGHT_MEALS_ONE_SNACK = '8 meals + snack',
    NINE_MEALS_ONE_SNACK = '9 meals + snack',
    TEN_MEALS_ONE_SNACK = '10 meals + snack',
    TWO_MEALS_TWO_SNACKS = '2 meals + 2 snacks',
    THREE_MEALS_TWO_SNACKS = '3 meals + 2 snacks',
    FOUR_MEALS_TWO_SNACKS = '4 meals + 2 snacks',
    FIVE_MEALS_TWO_SNACKS = '5 meals + 2 snacks',
    SIX_MEALS_TWO_SNACKS = '6 meals + 2 snacks',
}